+++ frontend/src/components/tracks/TrackBanner.vue
