+++ frontend/src/components/tracks/TrackItem.vue
