+++ frontend/src/components/tracks/TrackListItem.vue
