+++ frontend/src/components/tracks/TrackParticipantItem.vue
