+++ frontend/src/store/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import authentication from './modules/authentication';
 import config from './modules/config';
-import podcasts from './modules/podcasts';
+import podcasts from './modules/tracks';
 import misc from './modules/misc';
 
 Vue.use(Vuex);
