--- frontend/src/store/modules/podcasts/index.js
@@ -1,12 +1,0 @@
-import state from './podcasts.state';
-import mutations from './podcasts.mutations';
-import actions from './podcasts.actions';
-import getters from './podcasts.getters';
-
-export default {
-  namespaced: true,
-  state,
-  mutations,
-  actions,
-  getters
-};
