+++ frontend/src/store/modules/tracks/index.js
@@ -1,0 +1,12 @@
+import state from './tracks.state';
+import mutations from './tracks.mutations';
+import actions from './tracks.actions';
+import getters from './tracks.getters';
+
+export default {
+  namespaced: true,
+  state,
+  mutations,
+  actions,
+  getters
+};
